Let Dependabot manage vmtest version updates
It's easy to forget to update the vmtest version used by the action when
a new vmtest release is cut. In an attempt to semi-automate the process,
use Dependabot to manage said version in a dummy Cargo manifest. It
will run daily and create a pull request if a new version is
encountered. In the action we then pick up the version managed in this
manifest when installing the vmtest binary.

Signed-off-by: Daniel Müller <deso@posteo.net>